MvvM AccountType buttons finished ADD, SEARCH, UPDATE & DELETE buttons

+++ StoreAccountingApp/Models/AccountTypeService.cs
@@ -4,12 +4,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 
 namespace StoreAccountingApp.Models
 {
     public class AccountTypeService
     {
         _DBStoreAccountingContext ctx;
+
+        public object DialogResult { get; private set; }
+
         public AccountTypeService()
         {
             ctx = new _DBStoreAccountingContext();
@@ -18,7 +22,7 @@ namespace StoreAccountingApp.Models
         public List<AccountTypeDTO>GetAll()
         {
             List<AccountTypeDTO> accountTypeList = new List<AccountTypeDTO>();
-            var ObjQuery = from AccountType in ctx.AccountTypes
+            var ObjQuery =  from AccountType in ctx.AccountTypes
                             select AccountType;
             foreach (var accountType in ObjQuery)
             {
@@ -29,16 +33,38 @@ namespace StoreAccountingApp.Models
         public bool Add(AccountTypeDTO newAccountType)
         {
             //bool IsAdded = false;
-
-            var objAccountType = new AccountType()
+            //                                                          <----- Add validations here
+            var ObjAccountTypeIdExist = ctx.AccountTypes.Find(newAccountType.AccountTypeId);
+            if (ObjAccountTypeIdExist != null)
             {
-                Name = newAccountType.Name,
-                Description = newAccountType.Description
-            };
+                MessageBoxResult dialogResult = MessageBox.Show(    $"An account type with id {newAccountType.AccountTypeId} was already found, do you want to update it instead?",
+                                                                "Accountype already exists", MessageBoxButton.YesNo);
+                if (dialogResult == MessageBoxResult.Yes)
+                    return Update(newAccountType);
+                else
+                    throw new ArgumentException($"Add operation failed, {newAccountType.AccountTypeId} already exists")
 
-            ctx.AccountTypes.Add(objAccountType);
-            //IsAdded = ctx.SaveChanges() > 0;
-            return ctx.SaveChanges() > 0; ;
+            }
+            var ObjAccountTypeNameExist = ctx.AccountTypes.FirstOrDefault(a => a.Name == newAccountType.Name);
+            if (ObjAccountTypeNameExist != null)
+                throw new ArgumentException($"Add operation failed, {newAccountType.Name} already exists");
+            try
+            {
+                var objAccountType = new AccountType()
+                {
+                    Name = newAccountType.Name,
+                    Description = newAccountType.Description
+                };
+                ctx.AccountTypes.Add(objAccountType);
+                //IsAdded = ctx.SaveChanges() > 0;
+                return ctx.SaveChanges() > 0; ;
+            }
+            catch (Exception)
+            {
+
+                throw;
+            }
+
         }
         public AccountTypeDTO Search(int accountTypeId)
         {
@@ -58,14 +84,18 @@ namespace StoreAccountingApp.Models
         public bool Update(AccountTypeDTO objAccountTypeToUpdate)
         {
             var ObjAccountType = ctx.AccountTypes.Find(objAccountTypeToUpdate.AccountTypeId);
-            ObjAccountType.Name = objAccountTypeToUpdate.Name;
-            ObjAccountType.Description = objAccountTypeToUpdate.Description;
+            if (ObjAccountType != null)
+            {
+                ObjAccountType.Name = objAccountTypeToUpdate.Name;
+                ObjAccountType.Description = objAccountTypeToUpdate.Description;
+            }
             return ctx.SaveChanges() > 0;
         }
         public bool Delete(int accountTypeId)
         {
             var ObjAccountTypeToDelete = ctx.AccountTypes.Find(accountTypeId);
-            ctx.AccountTypes.Remove(ObjAccountTypeToDelete);
+            if (ObjAccountTypeToDelete != null)
+                ctx.AccountTypes.Remove(ObjAccountTypeToDelete);
             return ctx.SaveChanges() > 0;
         }
     }
